Advanced Medical Review
This is a review of the loadouts to make sure that they have appropriate medical supplies for ADVM. GB loadouts have been left alone at Sams request.

I've had to tweak things here and there to make room for the extra supplies, this has been done in-line with the loadout theme.